inteligent read csv to reduce memory, reduce memory in scaler and decomposer, deactivate stack_Xs to reduce memory

+++ xenon_ext/decompose/flexible_decomposer.py
@@ -60,6 +60,6 @@ class FlexibleDecomposer(BaseEstimator, TransformerMixin):
             return pd.DataFrame(
                 Xt,
                 columns=[f"component-{i}" for i in range(Xt.shape[1])],
-                index=X.index)
+                index=X.index, dtype="float32")
         else:
             return Xt
